MarkedTextを別ファイルに分けた
